Refactor: View, ViewController 분리
