+++ 34th-Sopt-Assignment/34th-Sopt-Assignment.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "171d0819b80f28bef4fc21b264ae431284970bc01cfab64666be6ccbb80c93fd",
+  "originHash" : "81778f14f0763f15f4e7dbde3761edeab241e4dc7f7938a6296fb6156172ef81",
   "pins" : [
     {
       "identity" : "alamofire",
