+++ 34th-Sopt-Assignment/34th-Sopt-Assignment/Presentation/Main/Controllers/BaseViewController.swift
@@ -3,6 +3,7 @@ import SnapKit
 import Then
 
 final class BaseViewController: UIViewController {
+    
     // MARK: - Property
 //    private var titles = ["홈", "실시간", "TV프로그램", "영화", "파라마운트"]
     
@@ -12,6 +13,7 @@ final class BaseViewController: UIViewController {
 //    private var containerView: UIView!
 //    private var contentViewControllers: [UIViewController] = []
     private let tvingLogo = UIImageView()
+    
     private let profile = UIImageView()
     
     // MARK: - Life Cycle
@@ -29,6 +31,7 @@ final class BaseViewController: UIViewController {
             let image = UIImage(named: "profile")
             $0.image = image
         }
+        
         tvingLogo.do{
             let image = UIImage(named: "logo1")
             $0.image = image
@@ -52,7 +55,7 @@ final class BaseViewController: UIViewController {
         }
     }
     
-    func setViewHierarchy(){
+    func setViewHierarchy() {
         view.addSubviews(tvingLogo, profile)
     }
