+++ 34th-Sopt-Assignment/34th-Sopt-Assignment/Presentation/Main/Controllers/RecommendCollectionView.h
@@ -1,0 +1,12 @@
+//
+//  RecommendCollectionView.h
+//  34th-Sopt-Assignment
+//
+//  Created by Chandrala on 6/7/24.
+//
+
+#ifndef RecommendCollectionView_h
+#define RecommendCollectionView_h
+
+
+#endif /* RecommendCollectionView_h */
